+++ Cascade/ReturnCheck.cc
@@ -1,0 +1,5 @@
+#include "ReturnCheck.hh"
+
+namespace cascade {
+  check_return_map gReturnMap;
+}
